feat: introduce `depFilter` option
