+++ src/main/ts/topo.ts
@@ -1,0 +1,168 @@
+import glob from 'fast-glob'
+import { analyze } from 'toposource'
+import { dirname, join, relative, resolve } from 'path'
+import { promises as fs } from 'fs'
+
+import {
+  ITopoOptionsNormalized,
+  IPackageEntry,
+  IPackageJson,
+  ITopoOptions,
+  ITopoContext
+} from './interface'
+
+export * from './interface'
+
+export const getPackages = async (
+  options: ITopoOptionsNormalized
+): Promise<Record<string, IPackageEntry>> => {
+  const { pkgFilter } = options
+  const manifestsPaths = await getManifestsPaths(options)
+  const manifests: IPackageJson[] = await Promise.all(
+    manifestsPaths.map(p => readJsonFile(p))
+  )
+
+  checkDuplicates(manifests)
+
+  return manifests.reduce<Record<string, IPackageEntry>>((m, p, i) => {
+    const absPath = dirname(manifestsPaths[i])
+    const relPath = relative(options.cwd, absPath)
+    const entry = {
+      name: p.name,
+      manifest: p,
+      manifestPath: manifestsPaths[i],
+      path: relPath, // legacy
+      relPath,
+      absPath
+    }
+
+    if (pkgFilter(entry)) {
+      m[p.name] = entry
+    }
+
+    return m
+  }, {})
+}
+
+const checkDuplicates = (manifests: IPackageJson[]): void => {
+  const duplicates = manifests
+    .map(m => m.name)
+    .filter((e, i, a) => a.indexOf(e) !== i)
+  if (duplicates.length > 0) {
+    throw new Error(`Duplicated pkg names: ${duplicates.join(', ')}`)
+  }
+}
+
+export const getRootPackage = async (cwd: string): Promise<IPackageEntry> => {
+  const manifestPath = resolve(cwd, 'package.json')
+  const manifest = await readJsonFile(manifestPath)
+
+  return {
+    name: manifest.name,
+    manifest,
+    manifestPath,
+    path: '/',
+    relPath: '/',
+    absPath: dirname(manifestPath)
+  }
+}
+
+export const topo = async (
+  options: ITopoOptions = {}
+): Promise<ITopoContext> => {
+  const {
+    cwd = process.cwd(),
+    filter = _ => true,
+    pkgFilter = filter,
+    depFilter = _ => true
+  } = options
+  const root = await getRootPackage(cwd)
+  const _options: ITopoOptionsNormalized = {
+    cwd,
+    filter,
+    depFilter,
+    pkgFilter,
+    workspaces: options.workspaces || root.manifest.workspaces || []
+  }
+  const packages = await getPackages(_options)
+  const { edges, nodes } = getGraph(
+    Object.values(packages).map(p => p.manifest),
+    depFilter
+  )
+  const { queue, graphs, next, prev, sources } = analyze([
+    ...edges,
+    ...nodes.map<[string]>(n => [n])
+  ])
+
+  return {
+    nodes,
+    edges,
+    queue,
+    graphs,
+    sources,
+    prev,
+    next,
+    packages,
+    root
+  }
+}
+
+export const getGraph = (
+  manifests: IPackageJson[],
+  depFilter: ITopoOptionsNormalized['depFilter']
+): {
+  nodes: string[]
+  edges: [string, string][]
+} => {
+  const nodes = manifests.map(({ name }) => name).sort()
+  const edges = manifests
+    .reduce<[string, string][]>((edges, pkg) => {
+      Object.entries({
+        ...pkg.dependencies,
+        ...pkg.devDependencies,
+        ...pkg.optionalDependencies,
+        ...pkg.peerDependencies
+      }).forEach(
+        ([name, version]) =>
+          depFilter({ name, version }) &&
+          nodes.includes(name) &&
+          edges.push([name, pkg.name])
+      )
+
+      return edges
+    }, [])
+    .sort()
+
+  return {
+    edges,
+    nodes
+  }
+}
+
+export const getManifestsPaths = async ({
+  workspaces,
+  cwd
+}: ITopoOptionsNormalized) =>
+  await glob(
+    workspaces.map(w => slash(join(w, 'package.json'))),
+    {
+      cwd,
+      onlyFiles: true,
+      absolute: true
+    }
+  )
+
+const readJsonFile = async (filepath: string) =>
+  JSON.parse(await fs.readFile(filepath, 'utf8'))
+
+// https://github.com/sindresorhus/slash/blob/b5cdd12272f94cfc37c01ac9c2b4e22973e258e5/index.js#L1
+export const slash = (path: string): string => {
+  const isExtendedLengthPath = /^\\\\\?\\/.test(path)
+  const hasNonAscii = /[^\u0000-\u0080]+/.test(path) // eslint-disable-line no-control-regex
+
+  if (isExtendedLengthPath || hasNonAscii) {
+    return path
+  }
+
+  return path.replace(/\\/g, '/')
+}
