+++ src/test/ts/topo.ts
@@ -4,7 +4,7 @@ import { dirname, join, resolve } from 'path'
 import { fileURLToPath } from 'url'
 import micromatch from 'micromatch'
 
-import { getManifestsPaths, ITopoOptions, topo } from '../../main/ts/index'
+import { getManifestsPaths, ITopoOptions, topo } from '../../main/ts/topo'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const fixtures = resolve(__dirname, '../fixtures')
@@ -13,7 +13,13 @@ test('`getManifestsPaths` returns absolute package.json refs', async () => {
   const cwd = resolve(fixtures, 'regular-monorepo')
   const workspaces = ['packages/**/*/', '!packages/e']
   const result = (
-    await getManifestsPaths({ cwd, workspaces, filter: () => true })
+    await getManifestsPaths({
+      cwd,
+      workspaces,
+      filter: () => true,
+      depFilter: () => true,
+      pkgFilter: () => true
+    })
   ).sort()
   const expected = ['a', 'c', 'd/d/d'].map(f =>
     join(cwd, 'packages', f, 'package.json')
@@ -94,7 +100,7 @@ test('`topo` returns monorepo digest: release queue, deps graph, package manifes
   assert.equal(result, expected)
 })
 
-test('`topo` applies filter', async () => {
+test('`topo` applies filter/pkgFilter', async () => {
   const cwd = resolve(fixtures, 'regular-monorepo')
   const workspaces = ['packages/*']
   const filter: ITopoOptions['filter'] = ({ path }) =>
@@ -149,6 +155,84 @@ test('`topo` applies filter', async () => {
     ],
     next: new Map([['e', ['c']]]),
     prev: new Map([['c', ['e']]])
+  }
+
+  assert.equal(result, expected)
+})
+
+test('`topo` applies depFilter', async () => {
+  const cwd = resolve(fixtures, 'regular-monorepo')
+  const workspaces = ['packages/*']
+  const depFilter: ITopoOptions['depFilter'] = () => false
+  const result = await topo({ cwd, workspaces, depFilter })
+
+  const expected = {
+    queue: ['a', 'c', 'e'],
+    nodes: ['a', 'c', 'e'],
+    edges: [],
+    sources: ['a', 'c', 'e'],
+    packages: {
+      a: {
+        name: 'a',
+        manifest: {
+          name: 'a',
+          private: true
+        },
+        manifestPath: join(cwd, 'packages/a/package.json'),
+        path: 'packages/a',
+        relPath: 'packages/a',
+        absPath: resolve(cwd, 'packages/a')
+      },
+      c: {
+        name: 'c',
+        manifest: {
+          name: 'c',
+          dependencies: {
+            e: '*'
+          }
+        },
+        manifestPath: join(cwd, 'packages/c/package.json'),
+        path: 'packages/c',
+        relPath: 'packages/c',
+        absPath: resolve(cwd, 'packages/c')
+      },
+      e: {
+        name: 'e',
+        manifest: {
+          name: 'e'
+        },
+        manifestPath: join(cwd, 'packages/e/package.json'),
+        path: 'packages/e',
+        relPath: 'packages/e',
+        absPath: resolve(cwd, 'packages/e')
+      }
+    },
+    root: {
+      name: 'root',
+      manifest: {
+        name: 'root'
+      },
+      manifestPath: join(cwd, 'package.json'),
+      path: '/',
+      relPath: '/',
+      absPath: resolve(cwd)
+    },
+    graphs: [
+      {
+        nodes: new Set(['a']),
+        sources: ['a']
+      },
+      {
+        nodes: new Set(['c']),
+        sources: ['c']
+      },
+      {
+        nodes: new Set(['e']),
+        sources: ['e']
+      }
+    ],
+    next: new Map(),
+    prev: new Map()
   }
 
   assert.equal(result, expected)
